Move preprocess functions to library
